develop/server 브랜치 최신화 (#6)
* Remove unnecessity file.

* 새롭게 작성, 페이지 넘김 구현 및 여러 기능 그리고 디자인

* gui commit #3

1. UI 수정

* gui commit #3

1. UI 수정

* git commit #4
1. 로그인 창 UI 변경
   - 성욱이와 UI 회의 결과 많은 부분 수정하기로 함

2. 로고 일부 변경
   - 기존 로고에서 새로운 로고 도입
   - 추후 로고 수정 가능성 있음
