--- client/empty.txt
@@ -1,1 +1,0 @@
-이 파일을 삭제하고 GUI 프로젝트를 넣어주세요.
